+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unrelease]
 ### Added
 - added alternative dice notation
+- added outputs to result
 
 ## [0.0.8]
 ### Added